Restyle the New Review Request page and update for mobile devices.
The New Review Request page has been restyled to fit in with the new
look and feel of Review Board. It looks a bit like the dashboard now,
with a consistent look for the sidebar and similar color schemes
throughout the page.

The entire page works on both mobile and desktop screen sizes, allowing
people to post existing commits for review on a phone or tablet.

The search/filter field has been made more prominent. It's no longer
hidden behind a button click.

The RBTools instructions have been slightly changed to say why we
recommend using RBTools. It also links to the downloads page, instead of
the docs page.

The drag-and-drop area for uploading a diff no longer says you can
drag-and-drop if the browser/device doesn't support this.

The commit lists use icons instead of text for differentiating between
the revision, author, and time. This particularly helps on mobile
devices, and generally makes the list more scannable.

Any commit already posted for review now appears slightly faded out,
with a blueish circle arrow icon for jumping to the review request. This
stands out nicely, and is clearly clickable/tappable.

When using this page on a mobile device, it will start out showing the
list of repositories. Tapping one slides over the commits list. There's
a "< Repositories" link that takes the user back to the repositories
list.

Testing Done:
Tested on Chrome (Mac), Chrome (Nexus 7), Firefox, Safari (Mac), Safari (iOS).

I tested uploading changes and posting commits for review.

Tested going back and forward between the repos list and the commits list on mobile.

Tested filtering repositories on desktop and mobile.

Tested jumping to commits on desktop and mobile.

Reviewed at https://reviews.reviewboard.org/r/7269/