Integrate archive/mute with the dashboard.
This change adds new dashboard labels and bulk edit commands for archive and
mute.

The bulk commands work very much like the existing close commands, except I've
only added a prompt in the mute case, and I've skipped the "success" dialog
(it's obvious because the dashboard immediately reloads without those items or
with those items showing labels).

For the dashboard labels, I've done a little refactoring to use a list of
tuples when building the labels HTML, rather than a dictionary. This is both
somewhat more efficient, and the resulting order of items is more stable.

When testing the commands, I came to notice that it was possible to "unmute" an
archived review request or "unarchive" a muted one. Even though this isn't
critically bad, it was unexpected, so I fixed that up.

Testing done:
- Checked that the correct dashboard labels were shown everywhere.
- Checked that dashboard labels always appeared in the same order.
- Ran all the bulk commands and verified the results.
- Ran unit tests.
- Ran js-tests.

Reviewed at https://reviews.reviewboard.org/r/7287/