Hide the archive menu when logged out
Previously the unarchive and hide icon was shown to users that were not
logged in. The menu would not appear and clicking on the icon would not
do anything. Now we just hide the UI from users that are not logged in.

Testing Done:
No JS errors in browser console.

Manually verified the UI is still shown when logged in.

Reviewed at https://reviews.reviewboard.org/r/7286/