Only render visibility for authenticated users
Previously, when an unauthenticated user attempted to view the 'All
Review Requests' page, every summary would be `None`. This was because
the `SummaryColumn` would attempt to evalute the
`review_request.visibility` attribute, but it is only defined when the
user is authenticated. This would lead to an exception being raised in
the rendering and `None` being returned instead. This patch fixes the
issue by only attempting to render the review request visibility when
the user viewing the datagrid is authenticated.

Testing Done:
The archived and muted state is correctly shown.

The summaries are correctly rendered in the datagrid.

Reviewed at https://reviews.reviewboard.org/r/7302/