Only show history entry symbol in commit diffs
The history entry symbol (e.g., `+` or `-`) column is now only shown in
the `DiffCommitIndexView` when there is a commit history diff to show.
This results in less whitespace on the left hand side of the table and
the diff complexity icons will align with those in the
`DiffFileIndexView`.

Testing Done:
Verified the extra column is not shown when there is no commit diffing
being done, or when the two histories are identical.

Reviewed at https://reviews.reviewboard.org/r/7303/