Add a smart tooltip to show descriptions in the diffviewer
Hovering over a summary in the commit description table now shows a
tooltip with copyable text if the commit's description is longer than
he summary. Moving the mouse off of the tooltip causes it to disappear.

Testing Done:
Verified that the description was correctly shown.

The tooltip's text align's pixel-perfectly with the table's text in
Chrome, Firefox, and Safari on OS X.

Reviewed at https://reviews.reviewboard.org/r/7256/