"Trivial" publishes for review requests
Each time a review request is published, it will send an e-mail (assuming
e-mails are turned on for the server as a whole). We want to add a checkbox
with the label "Send E-Mail", in the green draft header. If the user un-checks
this box before publishing, it will publish without sending an e-mail.

Testing done:
- Tested with new diff, checked and unchecked for "Send E-Mail".
- Tested with attachment only draft, checked and unchecked for "Send E-Mail".

Fixes bug 2761.

Reviewed at https://reviews.reviewboard.org/r/7022/