Make the review request page mobile-friendly.
The review request page is one of the most important pages on
Review Board, and we want it to be fully usable on mobile devices. This
change gets us a huge step of the way there by making the page
mobile-friendly. It now fits on a typical mobile phone's screen. All
operations, such as closing review requests, replying to comments, etc.
are doable as well.

Most of the changes are pretty straight-forward. Horizontal content has
become more vertical (avatars are above review, timestamps no longer
aligned to the upper-right, the issue summary table contents are
multi-row cells, and so on).

Diffed comment thumbnails are still horizontal (this will be addressed
in a later change).

The review request action bar had to be consolidated a bit. Now, all the
actions are hidden under a "..." button. When clicked, this unfolds to
show all available actions as a new, wrapped action bar.
Implementation-wise, it's just a special action menu.

These changes affect the diff viewer and file review UIs as well.
However, those pages are not optimized yet for mobile, and will be
updated in future changes.

Testing Done:
Tested that the main desktop-oriented layout continued to look correct on
Chrome and Firefox.

Tested the mobile UI on the iPhone and in responsive mode at different
resolutions in Chrome.

Tested review request actions, replying to comments, expanding/collapsing
boxes, and toggling issue states.

Reviewed at https://reviews.reviewboard.org/r/7288/