Add utility to find commit intervals
A commit interval `(A, B]` contains all the commits after `A` that lead
to `B`. That is, they are all the descendants of `A` that are ancestors
of `B`. This can now be computed via the `get_commit_interval` utility
method. Unit tests have been added to test various cases.

Testing Done:
Ran unit tests.

Reviewed at https://reviews.reviewboard.org/r/7325/