Allow diff generation starting at a base commit
The diffviewer can now generate a commit starting at a certain commit.
This allows the diffs of individual commits in a commit history to be
viewed, as well as spans of commits.

This change also updates the WebAPI's `DiffContextResource` so that the
commit selection UI can request ranges of commits instead of the entire
history.

Testing Done:
Ran unit tests.

Reviewed at https://reviews.reviewboard.org/r/7322/