Add diff commit selection UI to diffviewer
The diffviewer can now select two commits in a commit history to
generate a diff between. This currently does not support inter-commit
diffs and the UI is disabled in that case.

Testing Done:
Manually tested commit generation between individual commits and for
spans of commits.

Manually verified the commit selection UI is not shown for interdiffs.

Reviewed at https://reviews.reviewboard.org/r/7323/