Fix the js-tests.
The tests for the RepositorySelectionView were broken because of a template
change that introduced a space into the text content of the nodes. I've changed
this back to remove the extra space character.

While I was editing this I saw that there was some now-unused code that hadn't
been removed after the refactor/redesign.

Testing done:
- Ran js-tests.
- Verified interaction with the search box on the New Review Request page.

Reviewed at https://reviews.reviewboard.org/r/7331/