Add a spy for retinaGravatar to fix up some warnings.
When running the js-tests, there are a bunch of console errors about gravatars.
This is happening because we don't have gravatar source images, and the
retinaGravatar code is trying to parse "undefined?s=32", which doesn't make a
lot of sense. Because we don't actually care about this, I've spied out the
retinaGravatar code and just escape hatched it.

Testing done:
Ran js-tests.

Reviewed at https://reviews.reviewboard.org/r/7335/