Patch credentials into the git paths when verifying repositories.
If an http(s)-based git repository needs credentials to access, but the user
doesn't want to put the credentials into the URL, it's not possible to add it.
The problem is that even though the repository isn't used to access files
(since we use the raw file URL mask for that), we still try to do `git
ls-remote` on it.

With this change, if the user added credentials to the admin UI, we'll now
patch those credentials into the URL passed into ls-remote.

Testing done:
Ran unit tests.

Reviewed at https://reviews.reviewboard.org/r/7356/