Update, move "Account Settings" docs
The "Account Settings" documentation was put in at the end of the first chapter
("Getting Started"), but it's likely that users won't care about customization
until they're more familiar with the tool generally. This change moves the
settings documentation to the very end of the manual, and updates it with
everything that's been added since we last updated it.

This also slips in a couple tiny formatting changes to the "What is Code
Review" page.

Testing done:
Built HTML and looked at the result. Checked hyperlinks.

Reviewed at https://reviews.reviewboard.org/r/7361/