Add a base foundation for Python codebase documentation.
This adds a new section to the Extending Review Board documentation that
links to auto-generated docs for the most interesting parts of the
codebase. These are the parts that are most likely to be useful to
extension writers and contributors looking to understand the main parts
of Review Board. Models, managers, registration functions, and so on.

The actual documentation files are auto-generated using Sphinx's
autosummary. These are not checked into source control. However, any new
module added to Review Board that requires documentation needs to go
into docs/manual/extending/coderef/index.rst.

Each module, class, and function in the docs link to the source code on
GitHub, taking into account the Git SHA and the branch/tag for the
current version.

Testing Done:
Built the docs and went through the modules. The content was there. There's
still bits and pieces to fix up in the docs, which will be done in separate
changes.

Reviewed at https://reviews.reviewboard.org/r/7363/