Add some documentation to our top-level modules.
This adds documentation to the `reviewboard`, `reviewboard.rb_platform`, and
`reviewboard.signals` modules. Most of these had docs already, but not in
a format that would be picked up by Sphinx.

The docstrings at the top of the files show up as summaries in the
generated list of modules. The full details show up on the given module
pages. As we go forward, we'll want to flesh out more module docstrings,
to help guide contributors and developers through the relevant parts of
the codebase.

Testing Done:
Built the docs. Saw the details on the module pages, and the summaries on
the module listings.

Reviewed at https://reviews.reviewboard.org/r/7367/