Fix the review request actions bar when empty.
When there was nothing to display in the review request actions bar,
it'd appear as a thin line, which looked horrible. This was because
there wasn't any content on the bar to force the height to anything
reasonable.

We already have a :before style, with no padding or content. This change
updates it to have some content (which is hidden but still takes up
room), and the same padding that the actions themselves have.

Also, I previously swapped the names of two variables (the ones use for
vert and horiz padding on actions). The vert was containing the horiz
padding, and vice-versa. These have been fixed up.

Testing Done:
Viewed an interdiff, which has no actions. Saw that the bar was the correct height.

Reviewed at https://reviews.reviewboard.org/r/7375/