Use "readonly" instead of "disabled" for "Install Key" input.
Firefox has a terrible, horrible, no good, very bad bug whereby a text input
which is `disabled` cannot be selected. This is especially problematic for our
install key field, which is a 40-character long hash which we want people to be
able to copy and paste.

It turns out that the "readonly" property works equally well for disabling
changes, but Firefox allows selection on those. This property is technically
HTML5 but is well-supported across all browsers.

Testing done:
Verified that "readonly" did the right thing in both Firefox and Chrome.

Reviewed at https://reviews.reviewboard.org/r/7380/