Fix the z-index of the mobile header bar and logo.
The z-index of the header bar and logo was set way too high. This was
due to an initial design for the mobile stuff that would place the
header bar always on top, and keep things below it. That's not the
implementation we ended up with, though. This change places the
z-indexes much further down.

Testing Done:
Tested all the cases I could think of for relative z-indexes. Things
looked more correct. It's still possible we'll find some corner case
somewhere, though.

Reviewed at https://reviews.reviewboard.org/r/7377/