Don't double-normalize the destination file path.
A recent change added some improved normalization of file paths, which
would take care of doing join of the basedir path (if any) and the
destination file path in a diff.

However, there was still a second case where the path would get
normalized in exactly the same way, later in the file processing. It was
expecting to do the normalization, and didn't account for the
possibility that it'd already be normalized.

Since we're now normalizing both paths earlier, at the same time,
there's no reason to normalize again later. This just removes that step.

Testing Done:
Unit tests passed. They failed before the fix, with the double path reported.

Verified the workaround of adding a leading slash with the customer.

Reviewed at https://reviews.reviewboard.org/r/7399/