Fix Git diffs containing tabs after the filename.
We found a case where a user had a generated Git diff that contained
tabs after the filename, as if more data would follow. Our parser wasn't
expecting this, and ended up including the tab along with the filename.

Now, any extra whitespace around the filename is stripped away, allowing
the files to be properly matched against the repository.

Testing Done:
The new unit test failed without the fix, and passed with it.

Reviewed at https://reviews.reviewboard.org/r/7397/