Do not use both cumulative and condensed to refer to cumulative diffs
This patch updates the diff comment model and the diff comment resource
to use the term `cumulative_diff` instead of `condensed_diff`. The
option was named `cumulative_diff` in the diff chunk generator and the
diff chunk generator, so this is for consistency.

Testing Done:
Ran unit tests.

Reviewed at https://reviews.reviewboard.org/r/7404/