Diffviewer now switches commit ranges with a single revision
Previously, the diffviewer would fail to show the correct diff when
there was only a single revision and the commit range changed. This
occured becuase when there was only a single revision, the way the
revision was being retrieved was causing an error. This resulted in
the page not being updated. This behaviour has now been rectified.

Testing Done:
Verified the correct commit diff was shown when the selected commit
range changed.

Reviewed at https://reviews.reviewboard.org/r/7403/