Return a real error message when trying to close drafts via the API.
In the case where an API request was trying to close as submitted a review
request that hadn't been published yet, we would crash with an unhandled
exception. The problem here is that we added validation to prevent people from
marking unpublished review requests as submitted, but failed to handle that in
the resource code.

We had similar handling in other resources, so I've just copied the logic over.

Testing done:
Issued a close via the dashboard and didn't see an exception on my devserver.

Reviewed at https://reviews.reviewboard.org/r/7406/