Fix a vulnerability in the review request update bubbles.
The update bubbles were assuming that the URL and user's full name were
HTML-safe. While the former may have been true, the latter certainly
wasn't. This could be used to inject scripts into update bubbles.

We now properly escape the content, ensuring it's rendered as entered.

Testing Done:
Triggered the bubble without the fix and with a script injection in my
name. Saw that the script executed.

After the fix, the script appeared as plain text in the bubble.

Reviewed at https://reviews.reviewboard.org/r/7408/