Fix WebAPI errors returning empty messages
Previously, the WebAPI would sometimes return a empty error. This was
because we were referring to the `msg` attribute of the underlying
exception, instead of casting it to a text type.

Also fix some typos that were introduced in the errors module.

Unit tests have been introduced to test the behaviour of raising a
`ReopenError`, `CloseError`, and `PublishingError` during a review
request reopen, close, and publish, respectively.

Testing Done:
Ran unit tests.

Reviewed at https://reviews.reviewboard.org/r/7415/