Fix display of filediffs which are copied with no modifications.
In the case where a file had no modifications and was moved we wouldn't
bother trying to generate chunks. We now exhibit the same behaviour for
copied files.

In particular this fixes a bug with hg repositories where unmodified
copies would fail to display with an error about diff application.
This error does not seem to happen in the git case because of a
difference in how UNKNOWN revisions and diff application are handled.

As a follow up it might be worthwhile to investigate why unknown
revisions are represented and handled differently, but this fix is
still valid as we shouldn't be generating chunks in this case.

Testing Done:
Observed that the unmodified copy case is now fixed for hg, and remains the same for git.

Unit tests all pass.

Reviewed at https://reviews.reviewboard.org/r/7416/