Make the RB logo a link
The initial change for responsiveness removed the link on the Review
Board logo to use it as a button. We now no longer use it as a button
on mobile so this patch returns the logo to being a link to the root
page.

Testing Done:
Tried the link.

Reviewed at https://reviews.reviewboard.org/r/7434/