Add some structure to the comment dialog for extensibility.
The comment dialog's layout and structure was built for the existing set
of checkboxes and other UI, but with no flexibility for laying out
additional elements. One of the big problems is that, to properly size
the text field, we needed to account for the buttons and other content
below it, but this was hard-coding the checkboxes.

Now, the dialog has been structured for a header, body area, optional
widgets (which we can later allow extensions to add to), status, and
buttons. The layout code takes these main areas into account, rather
than the individual components.

Testing Done:
Tested in Chrome and Firefox, with the following tests:

* Default state for a new comment.
* Defautl state for an existing comment.
* Default state for a comment with other comments shown on the side.
* Toggling the dirty state with the message shown.
* Toggling the Markdown mode with the link visibility.
* Resizing under these different cases.

Reviewed at https://reviews.reviewboard.org/r/7441/