Don't crash on SVN commits that are missing a date.
It's valid for SVN commits not to have a date, and this will definitely
happen if ACLs prevent a user from accessing the commits. We now
properly handle the case where the date is missing, returning the commit
with whatever info we have.

This can result in a bunch of empty-looking commit entries in the New
Review Request page, but that's better than the current behavior of
crashing. We have no way to search for accessible commits, and it's too
expensive to try to keep searching for accessible commits, so this is
the best compromise we have for now.

Testing Done:
Unit tests pass.

Reviewed at https://reviews.reviewboard.org/r/7445/