Fix the New Review Request page with missing commit data.
The New Review Request page's commit list was expecting a valid date for
each commit, and crashed if one wasn't provided. This would occur if dealing
with an inaccessible commit on Subversion.

We now set a flag if the commit is inaccessible, and change our presentation
for that commit. The commit will say "Revision XYZ (not accessible on this
repository)", and won't be available to post for review.

Testing Done:
Hard-coded all the data for a commit to being blank, and tried viewing the
commit list. Saw that the data, while useless, didn't look flat-out broken.

Tried posting inaccessible commits for review, and couldn't. I could still
post accessible commits.

Reviewed at https://reviews.reviewboard.org/r/7446/