Fix up problems with text fields in the review dialog.
There were a few bugs present with body-top/body-bottom and comments in the
header view. The fundamental issue seems to be that the raw value wasn't being
set correctly, due to a couple different issues (we often had the Review object
loaded but not with the right text types included, and then the `data` field
wasn't properly plumbed in the `ready` method).

There was also a problem in comments where we'd save the `extraData` field but
not specify the specific attrs, which caused a `<p>` bug.

While I was in here I cleaned up some obvious style and code-duplication issues.

Testing done:
- Repeatedly tried adding a review, editing the header a few times, and then
  publishing the review. Previously I could very often cause the header to be
  surrounded in a `<p>` tag, and now I can't make it happen at all.
- Ran unit tests
- Ran js-tests

Reviewed at https://reviews.reviewboard.org/r/7449/