Allow for trivial review reply publishes
Trivial review replies can be published, that is, published without
sending an e-mail notification.

This patch also fixes up some of the semantics in the review request
publishing. The `not-trivial` id has been changed to a `send-email`
class as it makes a bit more semantic sense.

NB: This is heavily based on Jessica Qian's work in
https://reviews.reviewboard.org/r/7125/

Testing Done:
Ran unit tests. Ran JS tests.

Tested both review reply and review request publishing with and without
the "Send E-Mail" box checked. It works as expected.

Tested that the "Send E-Mail" checkboxes do not appear when e-mails are
disabled for the server.

Reviewed at https://reviews.reviewboard.org/r/7447/