Simplify the display of file attachment comments a bit.
Most of the file attachment stuff has had a visual revamp. This change cleans
up the last vestiges of stuff in the review dialog and review boxes: namely, we
don't care so much about displaying both the filename and the caption, in favor
of a unified "display name" for the attachment. We've also deprecated the file
type icons just about everywhere, so this removes it from the review box too.

This also removes the now-unused file_attachment_thumbnail.html template.

Testing done:
- Checked the display of file attachment comments everywhere and saw that they
  looked correct.
- Ran unit tests.
- Ran js-tests.

Reviewed at https://reviews.reviewboard.org/r/7456/