Fix incorrectly interpreting newly-saved text as HTML in replies.
When saving a reply, the text and then the richText properties would be
set. The view listened to both, and updated the rendered text each time
one was emitted. This led to a brief moment where the newly-saved text
was being interpreted incorrectly, right before being set properly.

If entering a script tag, this would execute the script. However, this
would only ever happen locally, during the save process, and would not
affect other users, nor would it execute when reloading the page.

We now have a signal used for updating the text, which we can listen to
instead, in order to update the rendered text only once.

Testing Done:
Replied with a script tag, and saw that it rendered correctly, without
executing the script.

Reviewed at https://reviews.reviewboard.org/r/7451/