Reset rich text of review request fields when importing from commits.
In our original version of our Markdown support, we had a global
rich_text flag for an entire review request. Because of that, we needed
to handle escaping text coming from commits or change descriptions (when
updating from those), in order to prevent the imported text from
rendering as Markdown if the review request was set to rich text mode.

We don't have this limitation anymore, and have had reports of
frustration where a commit message expected to be plain text ended up
being escaped.

We now reset the flags instead for any field we import, triggering the
expected behavior. The text will still appear escaped for edit if the
user has that option enabled, but won't appear escaped otherwise.

Testing Done:
Unit tests were updated to check for the different code paths for this.
They pass.

Bugs closed: 3898

Reviewed at https://reviews.reviewboard.org/r/7452/