Fix assertRaisesMessage calls on Python 2.7.9/2.7.10.
assertRaisesMessage on Python 2.7.9/2.7.10 has a known bug where it
can't be used as a context manager. This fixes the code to use the
old-style invocation, where a lambda is passed in, allowing the tests to
run again.

Testing Done:
Unit tests now pass.

Reviewed at https://reviews.reviewboard.org/r/7453/