Improve the diff chunk selector.
The diff chunk selector is now cleaner, both visually and code-wise.

The big borders are gone. Instead, we have highlighting on the side,
which no longer has a "jumpy" effect like the borders have. They also
don't have the offset problems, nor do they cover up content.

The code is a bit cleaner. This class is no longer a singleton, and
instead is owned by the page. Recalculating the positions and sizes is
less expensive, and done as infrequently as possible.

The result is a lighter-weight, less noisy chunk selector.

Testing Done:
Played around with the selector on different browsers. The results were
consistent.

Jumped all around the page, back-and-forth, with arrow keys and with clicking.

Tried zooming the page in/out, and resizing, while watching console.log
output. We calculated infrequently.

Reviewed at https://reviews.reviewboard.org/r/7454/