Catch unexpected errors when validating diffs.
Unexpected errors could cause the diff validation resource to return a
HTTP 500, which is not very useful. We now catch this and return a more
suitable error, and log a traceback.

Testing Done:
Generated a fake exception, and saw the error information logged and
returned through the API.

Had a customer test this as well.

Reviewed at https://reviews.reviewboard.org/r/7458/