Fix support for trusting unverified SVN hosts.
We've had complaints for a little while now about users being unable to
trust unverified SSL certificates for SVN hosts. This turned out to be a
regression due to an error normalization function raising errors instead
of returning them.

The function has been fixed, and it's now possible to one-click verify
an unverified host.

Testing Done:
Set up a SVN server with a self-signed cert. Added it, and instead of seeing
an error message, I was given information on the host with the ability to trust
it.

Bugs closed: 3888

Reviewed at https://reviews.reviewboard.org/r/7462/