Provide a sane error if encountering multiple conflicting repositories.
If two repositories conflict (one has a Mirror Path matching the Path of
another repository), then validating a diff or creating a review request
in the API will result in an HTTP 500 error, due to a
`MultipleObjectsReturned` exception.

We now check for this and provide a more reasonable error message
instead, with a suggestion for how to resolve the issue.

Testing Done:
Unit tests pass.

Reviewed at https://reviews.reviewboard.org/r/7459/