Fix a crash with the Changes API when unable to load a field type.
If a change was made to a review request with a pluggable field, and
that field was no longer available (due to the extension no longer being
enabled), the code would fail, resulting in an HTTP 500.

We now check to make sure the field is there before including the
serialized contents in the payload.

Testing Done:
I was hitting this problem locally. This fixed it.

Reviewed at https://reviews.reviewboard.org/r/7463/