Make all review request banners flush with page sides
The submitted and discarded review request banners are now flush with
the sides of the page like the draft banner. This also fixes the
vertical spacing between the header and actions line and the change
description label.

Testing Done:
Submitted and discarded a review request. Viewed the banners.

Reviewed at https://reviews.reviewboard.org/r/7485/