Move publishError events out of views to models. Clarify PublishError message.
This causes to error handlers for the "Ship It!" button and the Review Dialog
to be triggered.

Testing Done:
Simulated a PublishError on the server and verified that alerts popped up after
clicking the "Ship It!" button, the "Publish Review" button in the Review
dialog, and the "Publish" button in the draft banner for both a review and a
reply.

Reviewed at https://reviews.reviewboard.org/r/7479/