Fix layout sizing regressions with the review request page.
The revamp of the review request page broke some of the layout
calculations, resulting in the bottom of the review request box and the
bottom-most text field jumping just a bit on page load and on edit,
sometimes.

The addition/move of some margins/paddings ended up affecting these
calculations, which weren't updated to take the changes into account. The
calculations were also thrown off by the addition of a margin-top on the
"extra" box (where file attachments and stuff live), which affected
things differently on Chrome and Firefox.

Now, layout should work just like how it used to, before the UI update.

Testing Done:
Tested in Chrome and Firefox. Didn't see any issues on page load or when
going in/out of edit mode with any of the review requests I had working as
repro cases.

We'll see how this performs in the wild.

Reviewed at https://reviews.reviewboard.org/r/7514/