Add documentation for the new webhook resource.
This change adds the pieces needed to generate docs for the new webhook
resource. Specifically, it adds new doc entries for webhook-list and webhook,
links in the index, and a fixture to auto-generate examples from.

While I was doing this, I encountered lots of trouble because the types for the
`apply_to`, `encoding`, and `type` fields weren't correct (they were lists of
strings instead of tuples). This was causing the docs to fail to build in a
cryptic way as the `webapi-resource` sphinx class tried to import "a.ll"

Testing done:
Built docs and looked at the results.

Reviewed at https://reviews.reviewboard.org/r/7524/