Polish and improve the usability of the issue summary table.
The issue summary table has been reworked to be easier to work with and
more visually appealing.

The old table had borders between each row, which was just added noise.
It also was pretty text-heavy for status information, both in the
columns and in the counts (which themselves were off to the side, and
less useful there), and used select boxes for status filters (which
blended together too much and likely weren't well-used).

The new table has a cleaner, friendlier look to it. The borders between
rows are gone, except when differentiating between rows of different
status types (like Open vs. Resolved, when in the All Statuses view).

Each row has an icon showing the status type for that row, which is
easier to see at a glance than the status text column. That old column
is gone now, giving more room for the summary.

The timestamp column no longer shows absolute timestamps, but rather
relative ones, making it easier to see how old a comment is.

The status filter and counts have been merged. There's now a set of tabs
for choosing the view of the table. These use icons for Open, Resolved,
and Dropped, which match the rows in the table and the issue areas in
comments. This makes it much easier to differentiate counts and to
switch between the views.

The user filter is now shown directly above the user column, helping tie
things together nicely. The enties in the column also now show
gravatars, to help visually distinguish the users.

There is one side-effect of the change, which is that the API for
comments has changed from listing timestamps in a localized format to
listing them in ISO 8601 format. This was the only API using a
non-standard format for timestamps, and it was due to the issue summary
tracker. We no longer need the localized format, and reverted the API to
something sane.

Testing Done:
Python and JavaScript unit tests pass.

All tests were performed in Chrome, Firefox, and IE11 in both desktop and mobile modes:

* Tested filtering issues using tabs and user filter.
* Tested that the borders were only shown between visible items of differing states.
* Toggled issue statuses and saw that the table and counters were properly updated, along with the timestamp.
* Tested jumping to issues when clicking anywhere on them.
* Tested that clicking the name of the user navigated to the user page.
* Tested that resizing the window updated the position of the users filter.
* Tested that, if no issues are shown by default (due to being all resolved), the users filter stays at a default position docked to the very right.

Reviewed at https://reviews.reviewboard.org/r/7515/